removed second init for outputs
initially, this was intended to signal to the eye calibrator that the
eye position is unknown and potentially out of range. It turns out that
the calibrator does not care about whether or not we do this here and
still signals eye position to be at center before the first real update.